refactor routers
