+++ apps/web/client/src/server/api/routers/index.ts
@@ -1,14 +1,7 @@
-export * from './canvas';
-export * from './chat';
 export * from './code';
 export * from './domain';
 export * from './forward';
-export * from './frame';
 export * from './github';
-export * from './invitation';
-export * from './member';
 export * from './project';
-export * from './sandbox';
 export * from './subscription';
 export * from './user';
-export * from './user-canvas';
