+++ apps/web/client/src/server/api/routers/user/index.ts
@@ -1,0 +1,2 @@
+export * from './user';
+export * from './user-canvas';
